refactor: 공연 캘린더 fetch 로직을 자식에서 부모로 이동
