+++ src/components/pages/PerformanceCalendar/PerformanceCalendarPage.tsx
@@ -1,38 +1,33 @@
 'use client';
 
 import { useState, useMemo, useRef, useEffect } from 'react';
+import { useQuery } from '@tanstack/react-query';
+import { endOfMonth, format, startOfMonth } from 'date-fns';
 import { ArrowUp } from 'lucide-react';
+import LoadingOverlay from '@/components/common/LoadingOverlay/LoadingOverlay';
+import { performancesApi } from '@/services/performancesService';
 import { Performance } from '@/types/performance';
 import CalendarFilter from './CalendarFilter';
 import PerformanceCalendar from './PerformanceCalendar';
 import SelectedDatePerformances from './SelectedDatePerformances';
 
 const PerformanceCalendarPage = () => {
-  const [allPerformances, setAllPerformances] = useState<Performance[]>([]);
   const [filterValues, setFilterValues] = useState<{ visit?: string }>({});
   const [selectedDate, setSelectedDate] = useState<Date>(() => new Date());
+  const [currentMonth, setCurrentMonth] = useState(() => new Date());
   const [showScrollToTop, setShowScrollToTop] = useState(false);
   const detailRef = useRef<HTMLDivElement>(null);
 
-  useEffect(() => {
-    const handleScroll = () => {
-      const scrollY = window.scrollY;
-      setShowScrollToTop(scrollY > 300);
-    };
+  const startDate = format(startOfMonth(currentMonth), 'yyyy-MM-dd');
+  const endDate = format(endOfMonth(currentMonth), 'yyyy-MM-dd');
 
-    window.addEventListener('scroll', handleScroll);
-    return () => window.removeEventListener('scroll', handleScroll);
-  }, []);
-
-  useEffect(() => {
-    const handleScroll = () => {
-      const scrollY = window.scrollY;
-      setShowScrollToTop(scrollY > 300);
-    };
-
-    window.addEventListener('scroll', handleScroll);
-    return () => window.removeEventListener('scroll', handleScroll);
-  }, []);
+  const { data: allPerformances, isLoading } = useQuery({
+    queryKey: ['performances', startDate, endDate],
+    queryFn: () =>
+      performancesApi.getPerformances({ startDate, endDate, size: 100 }),
+    select: (res) => res.data.data ?? [],
+    staleTime: 1000 * 60 * 60,
+  });
 
   useEffect(() => {
     const handleScroll = () => {
@@ -46,6 +41,7 @@ const PerformanceCalendarPage = () => {
 
   const filteredPerformances = useMemo(() => {
     const { visit } = filterValues;
+    if (!allPerformances) return [];
 
     return allPerformances.filter((perf) => {
       const matchVisit = !visit || perf.visit === visit;
@@ -69,13 +65,17 @@ const PerformanceCalendarPage = () => {
   return (
     <div className='mx-auto max-w-2xl px-4 py-8'>
       <CalendarFilter onChange={setFilterValues} />
-      <PerformanceCalendar
-        performances={filteredPerformances}
-        selectedDate={selectedDate}
-        onSelectedDateChange={setSelectedDate}
-        onPerformancesFetched={setAllPerformances}
-        onDateClick={handleDateClick}
-      />
+      <div className='relative'>
+        {isLoading && <LoadingOverlay />}
+        <PerformanceCalendar
+          performances={filteredPerformances}
+          selectedDate={selectedDate}
+          onSelectedDateChange={setSelectedDate}
+          onDateClick={handleDateClick}
+          currentMonth={currentMonth}
+          onMonthChange={setCurrentMonth}
+        />
+      </div>
       <div ref={detailRef}>
         <SelectedDatePerformances
           date={selectedDate}
